Refactor Component Interfaces
