+++ pkg/README.md
@@ -1,3 +1,3 @@
 # `/pkg`
 
-Contains common interfaces and types that are expected to be imported by third-party components.
+Contains common interfaces and component that are expected to be imported by third-party components.
