+++ pkg/component/payload.go
@@ -1,4 +1,4 @@
-package types
+package component
 
 import (
 	"io"
@@ -13,18 +13,15 @@ type ImageData map[string]interface{}
 // DecodedPayload holds the Image bytes and accompanying Data
 type DecodedPayload struct {
 	Image interface{}
-	ImageData
 }
 
 // InputPayload holds a reader to image bytes and accompanying Data.
 type InputPayload struct {
 	io.Reader
 	ImageBytes
-	ImageData
 }
 
 type OutputPayload struct {
 	io.WriteCloser
 	ImageBytes
-	ImageData
 }
