+++ pkg/component/transaction.go
@@ -1,10 +1,13 @@
-package types
+package component
 
 // Transaction represent a transaction containing a streamable payload (the message) and a response channel,
 // which is used to indicate whether the payload was successfully processed and propagated to the next destinations.
 type Transaction struct {
 	// InputPayload is the message payload of this transaction.
 	InputPayload
+
+	// ImageData is the message data of this transaction.
+	ImageData
 
 	// ResponseChan should receive a response at the end of a transaction,
 	// The response itself indicates whether the payload was successfully processed and propagated
