--- pkg/types/package.go
@@ -1,3 +1,0 @@
-// Package types defines any general structs and interfaces used throughout the
-// code base.
-package types
